+++ src/main/kotlin/icu/windea/pls/ep/config/CwtConfigPostProcessors.kt
@@ -9,7 +9,6 @@ import icu.windea.pls.config.config.optionData
 import icu.windea.pls.config.config.tagType
 import icu.windea.pls.config.util.CwtConfigResolverManager
 import icu.windea.pls.config.util.manipulators.CwtConfigManipulator
-import icu.windea.pls.core.withRecursionGuard
 
 class CwtBaseConfigPostProcessor : CwtConfigPostProcessor {
     override fun postProcess(config: CwtMemberConfig<*>) {
@@ -43,6 +42,7 @@ class CwtInjectConfigPostProcessor : CwtConfigPostProcessor {
 
     override fun postProcess(config: CwtMemberConfig<*>) {
         val pathExpression = config.optionData { inject } ?: return
+
         val configsToInject = CwtConfigResolverManager.findConfigsByPathExpression(config.configGroup, pathExpression)
         if (configsToInject == null) {
             invalidPathExpression(pathExpression, config)
@@ -52,34 +52,33 @@ class CwtInjectConfigPostProcessor : CwtConfigPostProcessor {
             noMatched(pathExpression, config)
         }
 
-        withRecursionGuard {
-            withRecursionCheck("cwt-inject@${pathExpression}@${config.pointer}") a@{
-                val targetConfig = config
-                val originalConfigs = targetConfig.configs ?: return@a null
-                val injectedConfigs = CwtConfigManipulator.createListForDeepCopy()
-                configsToInject.forEach { configToInject ->
-                    injectedConfigs += deepCopyForInjection(configToInject, targetConfig)
-                }
-                if (injectedConfigs.isEmpty()) {
-                    noMatched(pathExpression, config)
-                    return@a null
-                }
-
-                val newConfigs = CwtConfigManipulator.createListForDeepCopy()
-                newConfigs += originalConfigs
-                newConfigs += injectedConfigs
-
-                val updated = updateChildConfigs(targetConfig, newConfigs)
-                if (!updated) {
-                    logger.warn("Config injection ignored because config cannot carry child configs (config: ${config})")
-                    return@a null
-                }
-                logger.info("Applied config injection (path expression: ${pathExpression}, config: ${config})")
-                null
-            }
-        } ?: run {
+        // avoid (shallow) recursion injection: if configs to inject also contain inject option, ignore directly
+        if (configsToInject.any { it.optionData { inject } != null }) {
             recursive(pathExpression, config)
+            return
         }
+
+        val targetConfig = config
+        val originalConfigs = targetConfig.configs ?: return
+        val injectedConfigs = CwtConfigManipulator.createListForDeepCopy()
+        configsToInject.forEach { configToInject ->
+            injectedConfigs += deepCopyForInjection(configToInject, targetConfig)
+        }
+        if (injectedConfigs.isEmpty()) {
+            noMatched(pathExpression, config)
+            return
+        }
+
+        val newConfigs = CwtConfigManipulator.createListForDeepCopy()
+        newConfigs += originalConfigs
+        newConfigs += injectedConfigs
+
+        val updated = updateChildConfigs(targetConfig, newConfigs)
+        if (!updated) {
+            logger.warn("Config injection ignored because config cannot carry child configs (config: ${config})")
+            return
+        }
+        logger.info("Applied config injection (path expression: ${pathExpression}, config: ${config})")
     }
 
     private fun invalidPathExpression(pathExpression: String, config: CwtMemberConfig<*>) {